+++ src/main/java/org/truffleruby/core/exception/ExceptionNodes.java
@@ -74,9 +74,54 @@ public abstract class ExceptionNodes {
             return self;
         }
 
-
-        @Specialization(guards = { "self != from", "isRubyException(from)" })
+        @Specialization(
+                guards = { "self != from", "isRubyException(from)", "!isNameError(from)", "!isSystemCallError(from)" })
         protected Object initializeCopy(DynamicObject self, DynamicObject from) {
+            initializeExceptionCopy(self, from);
+            return self;
+        }
+
+        @Specialization(guards = { "self != from", "isSystemCallError(from)" })
+        protected Object initializeSystemCallErrorCopy(DynamicObject self, DynamicObject from) {
+            initializeExceptionCopy(self, from);
+            Layouts.SYSTEM_CALL_ERROR.setErrno(self, Layouts.SYSTEM_CALL_ERROR.getErrno(from));
+            return self;
+        }
+
+        @Specialization(guards = { "self != from", "isNoMethodError(from)" })
+        protected Object initializeCopyNoMethodError(DynamicObject self, DynamicObject from) {
+            initializeExceptionCopy(self, from);
+            initializeNameErrorCopy(self, from);
+            Layouts.NO_METHOD_ERROR.setArgs(self, Layouts.NO_METHOD_ERROR.getArgs(from));
+            return self;
+        }
+
+        @Specialization(
+                guards = { "self != from", "isNameError(from)", "!isNoMethodError(from)" })
+        protected Object initializeCopyNameError(DynamicObject self, DynamicObject from) {
+            initializeExceptionCopy(self, from);
+            initializeNameErrorCopy(self, from);
+            return self;
+        }
+
+        protected boolean isNameError(DynamicObject object) {
+            return Layouts.NAME_ERROR.isNameError(object);
+        }
+
+        protected boolean isNoMethodError(DynamicObject object) {
+            return Layouts.NO_METHOD_ERROR.isNoMethodError(object);
+        }
+
+        protected boolean isSystemCallError(DynamicObject object) {
+            return Layouts.SYSTEM_CALL_ERROR.isSystemCallError(object);
+        }
+
+        private void initializeNameErrorCopy(DynamicObject self, DynamicObject from) {
+            Layouts.NAME_ERROR.setName(self, Layouts.NAME_ERROR.getName(from));
+            Layouts.NAME_ERROR.setReceiver(self, Layouts.NAME_ERROR.getReceiver(from));
+        }
+
+        private void initializeExceptionCopy(DynamicObject self, DynamicObject from) {
             Backtrace backtrace = Layouts.EXCEPTION.getBacktrace(from);
             if (backtrace != null) {
                 Layouts.EXCEPTION.setBacktrace(self, backtrace.copy(getContext(), self));
@@ -86,8 +131,6 @@ public abstract class ExceptionNodes {
             Layouts.EXCEPTION.setFormatter(self, Layouts.EXCEPTION.getFormatter(from));
             Layouts.EXCEPTION.setMessage(self, Layouts.EXCEPTION.getMessage(from));
             Layouts.EXCEPTION.setCause(self, Layouts.EXCEPTION.getCause(from));
-
-            return self;
         }
 
     }